requirements.txt
streamlit==1.42.2
supabase==2.11.0
pymilvus==2.5.4
langgraph==0.2.68
redis==5.2.1
python-dotenv==1.0.1
python-docx==1.1.2
pdfplumber==0.11.5
pypdf==5.8.0
pymupdf==1.25.1
pandas==2.2.3
openai==1.65.1
langchain_openai==0.3.6
langchain==0.3.13
sentence-transformers==3.4.1
pytesseract==0.3.13
Pillow==11.1.0
openpyxl==3.1.5
chromadb==0.6.3
langgraph==0.2.68
redis==5.2.1
